src/config/components/database.py
from dotenv import load_dotenv

from config.components.boilerplate import BASE_DIR

load_dotenv()


DATABASES = {
    "default": {
        "ENGINE": "django.db.backends.sqlite3",
        "NAME": BASE_DIR / "db.sqlite3",
    },
}

DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
