pyproject.toml
[project]
name = "gradus-test"
version = "1.0"
description = ""
authors = [
    {name = "Oleksandr Yanchar",email = "contact.oleksandryanchar@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "bleach (>=6.3.0,<7.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "daphne (>=4.2.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)"
]
